'make koji' skips branches that lack Koji build targets
The default operationg of 'make koji' now is to build for all
discovered branches that have Koji build targets.  Once a release is
EOL'ed, the targets in Koji go away but the code remains in dist-git.
Well, just skip those as we encounter them so that the default
operation of 'make koji' can be "build for everything".  Still allow
the $BRANCHES variable to override the set of branches to build.

+++ utils/submit-koji-builds.sh
@@ -120,6 +120,8 @@ cd "${CWD}" || exit
 cd "${WRKDIR}" || exit
 ${VENDORPKG} co "${PROJECT}"
 cd "${PROJECT}" || exit
+git config user.name "${GIT_USERNAME}"
+git config user.email "${GIT_USEREMAIL}"
 
 # Allow the calling environment to override the list of dist-git branches
 if [ -z "${BRANCHES}" ]; then
@@ -127,9 +129,14 @@ if [ -z "${BRANCHES}" ]; then
 fi
 
 for branch in ${BRANCHES} ; do
+    # skip this branch if there is no koji target
+    if ! koji list-targets | grep -q "${branch}" >/dev/null 2>&1 ; then
+        echo "*** skipping branch ${branch} because there are no Koji build targets"
+        continue
+    fi
+
+    # clean it
     git clean -d -x -f
-    git config user.name "${GIT_USERNAME}"
-    git config user.email "${GIT_USEREMAIL}"
 
     # make sure we are on the right branch
     ${VENDORPKG} switch-branch ${branch}
